add makefile and dockerized
