+++ README.md
@@ -1,1 +1,3 @@
 # kotlin-boilerplate
+
+Run docker
