+++ cli/build.gradle
@@ -4,8 +4,17 @@ buildscript {
     ext.jackson_version = "2.9.10"
 }
 
+plugins {
+    id 'application'
+}
+
+run {
+    args '8080'
+}
+
 dependencies {
-    implementation project(":core")
+    implementation project(path:":core")
+    implementation project(path: ':core', configuration: 'shadow')
     compile "io.ktor:ktor-server-netty:$ktor_version"
     compile "io.ktor:ktor-jackson:$ktor_version"
     compile "org.koin:koin-ktor:$koin_version"
@@ -15,3 +24,5 @@ dependencies {
     testCompile "io.ktor:ktor-server-test-host:$ktor_version"
     testCompile "org.koin:koin-test:$koin_version"
 }
+
+mainClassName = 'com.alice.boilerplate.MainKt'
