+++ cli/src/main/com/alice/boilerplate/heatlhz/Response.kt
@@ -1,0 +1,3 @@
+package com.alice.boilerplate.heatlhz
+
+data class Response(val status: String, val time: String)
