+++ cli/src/main/com/alice/boilerplate/injection/moduleApi.kt
@@ -1,0 +1,10 @@
+package com.alice.boilerplate.injection
+
+import com.alice.boilerplate.healthz.application.HealthService
+import com.alice.boilerplate.heatlhz.ApiHealth
+import org.koin.dsl.module
+
+val moduleApi = module {
+    single { HealthService() }
+    single { ApiHealth(get()) }
+}
