+++ cli/src/main/com/alice/boilerplate/router/routerService.kt
@@ -1,10 +1,10 @@
-package router
+package com.alice.boilerplate.router
 
+import com.alice.boilerplate.heatlhz.ApiHealth
 import com.fasterxml.jackson.core.util.DefaultIndenter
 import com.fasterxml.jackson.core.util.DefaultPrettyPrinter
 import com.fasterxml.jackson.databind.SerializationFeature
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
-import heatlhz.ApiHealth
 import io.ktor.application.Application
 import io.ktor.application.call
 import io.ktor.application.install
@@ -12,11 +12,11 @@ import io.ktor.features.CallLogging
 import io.ktor.features.Compression
 import io.ktor.features.ContentNegotiation
 import io.ktor.features.DefaultHeaders
+import io.ktor.jackson.jackson
 import io.ktor.response.respond
 import io.ktor.routing.get
 import io.ktor.routing.routing
 import org.koin.ktor.ext.inject
-import io.ktor.jackson.*
 
 fun Application.main() {
     // Install Ktor features
@@ -30,7 +30,7 @@ fun Application.main() {
                 indentArraysWith(DefaultPrettyPrinter.FixedSpaceIndenter.instance)
                 indentObjectsWith(DefaultIndenter("  ", "\n"))
             })
-            registerModule(JavaTimeModule())  // support java.time.* types
+            registerModule(JavaTimeModule()) // support java.time.* types
         }
     }
     // Lazy inject HealthService
