--- cli/src/main/heatlhz/Response.kt
@@ -1,4 +1,0 @@
-package heatlhz
-
-
-data class Response(val status: String, val time: String)
