--- cli/src/main/injection/moduleApi.kt
@@ -1,10 +1,0 @@
-package injection
-
-import healthz.application.HealthService
-import heatlhz.ApiHealth
-import org.koin.dsl.module
-
-val moduleApi = module {
-    single { HealthService() }
-    single { ApiHealth(get()) }
-}
