+++ cli/src/main/resources/application.conf
@@ -8,6 +8,6 @@ ktor {
   }
 
   application {
-    modules = [router.RouterServiceKt.main]
+    modules = [com.alice.boilerplate.router.RouterServiceKt.main]
   }
 }
